refactor: 重複コードを共通化 (refs #26, #27, #28)
CodeClimate の "similar-code" 指摘(#26/#27 index.js, #28 message-maker.js)
に対応し、現行コードに残る実重複を解消する。挙動は変えない。

- `WEEKDAYS_JA` / `formatJst` が message-maker.js と battle-message-maker.js
  にバイト同一で二重定義されていた(v3.0.0 統合時に発生) → 共通モジュール
  `date-format.js` に抽出し両者で import。
- index.js `salmonrun` 内、終了1時間前の追加お知らせ予約ブロックが3箇所
  反復していた → `scheduleSalmonExtraNote(endTime)` ヘルパーに抽出。

oxlint / oxfmt --check / node --check パス。実APIスモークテストで coop(金?判定)・
coop next・battle の出力が refactor 前と同一であることを確認。
